native: name every path that leaves a tic unresolved
unresolved was one bit: a tic either carried through or it did not, with
no way to tell why short of reading the SQL. unimplemented already names
its own paths this way; unresolved gets the same treatment.

sim::unresolved names thirty bits, one per condition that was already an
independent term in some OR chain feeding the column: a pickup with no
arm, a crossed special line, a damaging sector, a stuck damage call, a
stuck gunshot fold, a weapon cycle that hits an unwritten action or runs
past its own budget, an unfinished player move, a body frame that needs
an action, a used line no manual door handles, a locked door, a stuck
state cycle, a stuck chase, two movers standing too close for either
order to be safe, a missile or skull this does not move, a crossed
special a monster's own allow-list reaches, an unmodeled z-mover, more
than one melee attacker, a stuck attack routine, a kill that counts or
drops an item, a stuck fireball spawn, a melee draw count that could
still undercount, stuck missile thinking, two thinkers sharing a sector,
a stuck plane clip, a floor type that changes a sector's texture, a
plane that reverted, a stuck door thinker, and a crushing plane. A fold
whose own internal cause is not exposed as an independent SQL term at
the point unresolved reads it (the weapon cycle's four cases, the state
cycle's four, the chase's eight, missile thinking's nine) keeps its one
bit rather than being pulled apart, since splitting it means widening a
tuple other callers read too.

mask() folds a list of (bit, condition) pairs into a UInt64 with
bitOr, starting from whatever the tic already carries. Every site that
used to write toUInt8(a OR b OR c) now calls it, so a bit's own comment
is the only place its condition is written out.

unresolved widens to UInt64 in the schema and every reader: the driver's
Refusal, Session's polls, and every live test's own row. refusal.rs's
message now names the bits on both columns, the same "unresolved:
NAME, NAME" shape unimplemented already had.

Live proof, gametic 170 in demo3 (the imp's fireball, thrown the tic
before, is still going and neither the general movers' own stage nor
its Z-axis counterpart runs one for it, and its own momentum stands it
close enough to another momentum mover to crowd that too): the row
reads TX_CROWDED, TX_UNRUN, TZ_UNRUN, and the driver's message names
all three.

+++ driver/src/native/session.rs
@@ -165,7 +165,7 @@ struct FrameRow {
     /// The frame's own tic, read out of `native_frames` so the same query
     /// can look `native_state`'s refusal columns up by it.
     tic: u32,
-    unresolved: u8,
+    unresolved: u64,
     unimplemented: u64,
 }
 
@@ -174,7 +174,7 @@ struct FrameRow {
 #[derive(Row, Deserialize)]
 struct CommittedRow {
     tic: u32,
-    unresolved: u8,
+    unresolved: u64,
     unimplemented: u64,
 }
